מבצע-המאה · פאזה 3: ההסגר 39⇒7 — 31 שוחררו עם analyze-נקי ובדיקות ירוקות (המרה 98.9%)

+++ new/dart-maor/dup-field-value.dart
@@ -10,7 +10,7 @@ dynamic dupFieldValue(List fams, Map def, Map pick, Map edit) {
   if (edited != null) return edited;
   final picked = pick[key];
   // JS `??` נופל רק על null/undefined (0 עובר) ⇒ Dart `??` זהה.
-  final int idx = picked ?? fams.indexWhere((f) => _truthy(get(f)));
+  final int idx = ((picked ?? fams.indexWhere((f) => _truthy(get(f)))) as int);
   return get(fams[idx >= 0 ? idx : 0]);
 }
